requirements.txt
appdirs==1.4.4
attrs==19.3.0
autopep8==1.5.3
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pathspec==0.8.0
pdfminer.six==20200517
pycodestyle==2.6.0
pycryptodome==3.9.7
six==1.15.0
sortedcontainers==2.2.2
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
